ae4 con modificaciones

+++ PlaceMyBet/PlaceMyBet/Models/Mercado.cs
@@ -7,7 +7,7 @@ namespace PlaceMyBet.Models
 {
     public class Mercado
     {
-        public Mercado(int idMercado, int tipoMercado, double infocuotaOver, double infocuotaUnder, double dineroapostadoOver, double dineroapostadoUnder, int idEvento)
+        public Mercado(int idMercado, double tipoMercado, double infocuotaOver, double infocuotaUnder, double dineroapostadoOver, double dineroapostadoUnder, int idEvento)
         {
             this.idMercado = idMercado;
             this.tipoMercado = tipoMercado;
@@ -20,7 +20,7 @@ namespace PlaceMyBet.Models
 
         public int idMercado { get; set; }
 
-        public int tipoMercado { get; set; }
+        public double tipoMercado { get; set; }
         public double infocuotaOver { get; set; }
 
         public double infocuotaUnder { get; set; }
@@ -35,7 +35,7 @@ namespace PlaceMyBet.Models
 
     public class MercadoDTO
     {
-        public MercadoDTO(int tipoMercado, double infocuotaOver, double infocuotaUnder)
+        public MercadoDTO(double tipoMercado, double infocuotaOver, double infocuotaUnder)
         {
             this.tipoMercado = tipoMercado;
             this.infocuotaOver = infocuotaOver;
@@ -43,7 +43,7 @@ namespace PlaceMyBet.Models
             
         }
 
-        public int tipoMercado { get; set; }
+        public double tipoMercado { get; set; }
         public double infocuotaOver { get; set; }
 
         public double infocuotaUnder { get; set; }
